fastapi_service/requirements.txt
aio_pika
alembic
asyncpg
fastapi
httpx
psycopg2-binary
pydantic-settings
redis
sqlalchemy
uvicorn
